// bot/src/interfaces/metrics/MetricConfiguration.ts
// This file is a part of AlphaGameBot.
// 
//     AlphaGameBot - A Discord bot that's free and (hopefully) doesn't suck.
//     Copyright (C) 2025  Damien Boisvert (AlphaGameDeveloper)
// 
//     AlphaGameBot is free software: you can redistribute it and/or modify
//     it under the terms of the GNU General Public License as published by
//     the Free Software Foundation, either version 3 of the License, or
//     (at your option) any later version.
// 
//     AlphaGameBot is distributed in the hope that it will be useful,
//     but WITHOUT ANY WARRANTY; without even the implied warranty of
//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
//     GNU General Public License for more details.
// 
//     You should have received a copy of the GNU General Public License
//     along with AlphaGameBot.  If not, see <https://www.gnu.org/licenses/>.

import type { Counter, Gauge, Histogram } from "prom-client";

/**
 * Types of Prometheus metrics supported
 */
export enum PrometheusMetricType {
    GAUGE = "gauge",
    COUNTER = "counter",
    HISTOGRAM = "histogram"
}

/**
 * Configuration for a metric type
 */
export interface MetricConfiguration {
    /** Unique identifier for this metric */
    name: string;
    
    /** Human-readable description */
    description: string;
    
    /** Type of Prometheus metric */
    prometheusType: PrometheusMetricType;
    
    /** Prometheus metric name */
    prometheusName: string;
    
    /** Prometheus help text */
    prometheusHelp: string;
    
    /** Label names for Prometheus */
    prometheusLabels?: string[];
    
    /** Histogram buckets (only used if prometheusType is HISTOGRAM) */
    prometheusBuckets?: number[];
    
    /**
     * Function to process metric data and update Prometheus metric
     * @param metric The Prometheus metric instance (Gauge, Counter, or Histogram)
     * @param data The metric data to process
     */
    processData: (metric: Gauge | Counter | Histogram, data: unknown) => void;
}